Completable future task

+++ src/main/java/com/example/rest/Employee.java
@@ -3,12 +3,10 @@ package com.example.rest;
 import lombok.Builder;
 import lombok.Data;
 
-import java.math.BigInteger;
-
 @Data
 @Builder
 public class Employee {
     private String name;
     private Long id;
-    private BigInteger salary;
+    private Double salary;
 }
